requirements.txt
# Core LangChain packages (let pip resolve)
langchain>=0.1.0
langchain-community>=0.0.38
langchain-openai>=0.1.0
langchain-text-splitters>=1.1.1
langchain-huggingface>=0.1.2

# Vector stores and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# LLM API
openai>=1.30.0

# PDF parsing and FastAPI extras
pypdf
python-multipart

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Web
fastapi
uvicorn